reworking of mission view
